+++ Utilities/ocvalidate/ValidateUEFI.c
@@ -397,6 +397,7 @@ CheckUEFIOutput (
   UINT32               ErrorCount;
   OC_UEFI_CONFIG       *UserUefi;
   CONST CHAR8          *TextRenderer;
+  CONST CHAR8          *GopPassThrough;
   BOOLEAN              IsTextRendererSystem;
   BOOLEAN              IsClearScreenOnModeSwitchEnabled;
   BOOLEAN              IsIgnoreTextInGraphicsEnabled;
@@ -455,6 +456,14 @@ CheckUEFIOutput (
       DEBUG ((DEBUG_WARN, "UEFI->Output->SanitiseClearScreen is enabled on non-System TextRenderer (currently %a)!\n", TextRenderer));
       ++ErrorCount;
     }
+  }
+
+  GopPassThrough = OC_BLOB_GET (&UserUefi->Output.GopPassThrough);
+  if (AsciiStrCmp (GopPassThrough, "Enabled") != 0
+    && AsciiStrCmp (GopPassThrough, "Disabled") != 0
+    && AsciiStrCmp (GopPassThrough, "Apple") != 0) {
+    DEBUG ((DEBUG_WARN, "UEFI->Output->GopPassThrough is illegal (Can only be Enabled, Disabled, Apple)!\n"));
+    ++ErrorCount;
   }
 
   //